+++ tencentcloud/cwp/v20180228/models.go
@@ -6688,6 +6688,12 @@ type DescribeBashEventsRequest struct {
 	// 过滤条件。
 	// <li>Keywords - String - 是否必填：否 - 关键词(主机内网IP)</li>
 	Filters []*Filter `json:"Filters,omitempty" name:"Filters"`
+
+	// 排序方式：根据请求次数排序：asc-升序/desc-降序
+	Order *string `json:"Order,omitempty" name:"Order"`
+
+	// 排序字段：CreateTime-发生时间。ModifyTime-处理时间
+	By *string `json:"By,omitempty" name:"By"`
 }
 
 func (r *DescribeBashEventsRequest) ToJsonString() string {
@@ -6705,6 +6711,8 @@ func (r *DescribeBashEventsRequest) FromJsonString(s string) error {
 	delete(f, "Limit")
 	delete(f, "Offset")
 	delete(f, "Filters")
+	delete(f, "Order")
+	delete(f, "By")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeBashEventsRequest has unknown keys!", "")
 	}
@@ -6824,6 +6832,12 @@ type DescribeBruteAttackListRequest struct {
 	// <li>ModifyEndTime - String - 是否必填：否 - 最近攻击时间筛选，结束时间</li>
 	// <li>Banned - String - 是否必填：否 - 阻断状态筛选，多个用","分割：0-未阻断（全局ZK开关关闭），82-未阻断(非专业版)，83-未阻断(已加白名单)，1-已阻断，2-未阻断-程序异常，3-未阻断-内网攻击暂不支持阻断，4-未阻断-安平暂不支持阻断</li>
 	Filters []*Filter `json:"Filters,omitempty" name:"Filters"`
+
+	// 排序方式：根据请求次数排序：asc-升序/desc-降序
+	Order *string `json:"Order,omitempty" name:"Order"`
+
+	// 排序字段：CreateTime-首次攻击时间
+	By *string `json:"By,omitempty" name:"By"`
 }
 
 func (r *DescribeBruteAttackListRequest) ToJsonString() string {
@@ -6841,6 +6855,8 @@ func (r *DescribeBruteAttackListRequest) FromJsonString(s string) error {
 	delete(f, "Limit")
 	delete(f, "Offset")
 	delete(f, "Filters")
+	delete(f, "Order")
+	delete(f, "By")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeBruteAttackListRequest has unknown keys!", "")
 	}
@@ -7684,6 +7700,12 @@ type DescribeHostLoginListRequest struct {
 	// <li>Status - int - 是否必填：否 - 状态筛选1:正常登录；5：已加白,14:已处理，15：已忽略</li>
 	// <li>RiskLevel - int - 是否必填：否 - 状态筛选0:高危；1：可疑</li>
 	Filters []*Filter `json:"Filters,omitempty" name:"Filters"`
+
+	// 排序方式：根据请求次数排序：asc-升序/desc-降序
+	Order *string `json:"Order,omitempty" name:"Order"`
+
+	// 排序字段：LoginTime-发生时间
+	By *string `json:"By,omitempty" name:"By"`
 }
 
 func (r *DescribeHostLoginListRequest) ToJsonString() string {
@@ -7701,6 +7723,8 @@ func (r *DescribeHostLoginListRequest) FromJsonString(s string) error {
 	delete(f, "Limit")
 	delete(f, "Offset")
 	delete(f, "Filters")
+	delete(f, "Order")
+	delete(f, "By")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeHostLoginListRequest has unknown keys!", "")
 	}
@@ -9108,6 +9132,12 @@ type DescribePrivilegeEventsRequest struct {
 	// 过滤条件。
 	// <li>Keywords - String - 是否必填：否 - 关键词(主机IP)</li>
 	Filters []*Filter `json:"Filters,omitempty" name:"Filters"`
+
+	// 排序方式：根据请求次数排序：asc-升序/desc-降序
+	Order *string `json:"Order,omitempty" name:"Order"`
+
+	// 排序字段：CreateTime-发现时间
+	By *string `json:"By,omitempty" name:"By"`
 }
 
 func (r *DescribePrivilegeEventsRequest) ToJsonString() string {
@@ -9125,6 +9155,8 @@ func (r *DescribePrivilegeEventsRequest) FromJsonString(s string) error {
 	delete(f, "Limit")
 	delete(f, "Offset")
 	delete(f, "Filters")
+	delete(f, "Order")
+	delete(f, "By")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribePrivilegeEventsRequest has unknown keys!", "")
 	}
@@ -9602,6 +9634,12 @@ type DescribeReverseShellEventsRequest struct {
 	// 过滤条件。
 	// <li>Keywords - String - 是否必填：否 - 关键字(主机内网IP|进程名)</li>
 	Filters []*Filter `json:"Filters,omitempty" name:"Filters"`
+
+	// 排序方式：根据请求次数排序：asc-升序/desc-降序
+	Order *string `json:"Order,omitempty" name:"Order"`
+
+	// 排序字段：CreateTime-发生时间
+	By *string `json:"By,omitempty" name:"By"`
 }
 
 func (r *DescribeReverseShellEventsRequest) ToJsonString() string {
@@ -9619,6 +9657,8 @@ func (r *DescribeReverseShellEventsRequest) FromJsonString(s string) error {
 	delete(f, "Limit")
 	delete(f, "Offset")
 	delete(f, "Filters")
+	delete(f, "Order")
+	delete(f, "By")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "DescribeReverseShellEventsRequest has unknown keys!", "")
 	}
@@ -9732,7 +9772,7 @@ type DescribeRiskDnsListRequest struct {
 	// 排序方式：根据请求次数排序：asc-升序/desc-降序
 	Order *string `json:"Order,omitempty" name:"Order"`
 
-	// 排序字段：AccessCount-请求次数
+	// 排序字段：AccessCount-请求次数。MergeTime-最近请求时间
 	By *string `json:"By,omitempty" name:"By"`
 }