+++ tencentcloud/organization/v20210331/errors.go
@@ -98,6 +98,9 @@ const (
 	// 企业组织不存在。
 	RESOURCENOTFOUND_ORGANIZATIONNOTEXIST = "ResourceNotFound.OrganizationNotExist"
 
+	// 集团服务不存在。
+	RESOURCENOTFOUND_ORGANIZATIONSERVICENOTEXIST = "ResourceNotFound.OrganizationServiceNotExist"
+
 	// 操作不支持。
 	UNSUPPORTEDOPERATION = "UnsupportedOperation"