+++ tencentcloud/vod/v20180717/models.go
@@ -13422,7 +13422,7 @@ type PushUrlCacheRequest struct {
 	// 预热的 URL 列表，单次最多指定20个 URL。
 	Urls []*string `json:"Urls,omitempty" name:"Urls"`
 
-	// 点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。
+	// <b>点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。</b>
 	SubAppId *uint64 `json:"SubAppId,omitempty" name:"SubAppId"`
 }